function2/requirements.txt
functions-framework==3.*
google-cloud-bigquery==2.33.0 
google-cloud-storage==2.3.0  